+++ MathPhysicEngine/Rigidbody/Matrix33.h
@@ -5,14 +5,14 @@ class Vector3D;
 
 class Matrix33 {
 private:
-	float values[9];
+	double values[9];
 
 public:
 
 	Matrix33(
-		float a = 0, float b = 0, float c = 0, 
-		float d = 0, float e = 0, float f = 0, 
-		float g = 0, float h = 0, float i = 0) :values{ a,b,c,d,e,f,g,h,i } {};
+		double a = 0, double b = 0, double c = 0, 
+		double d = 0, double e = 0, double f = 0, 
+		double g = 0, double h = 0, double i = 0) :values{ a,b,c,d,e,f,g,h,i } {};
 
 	Matrix33 operator*(const Matrix33& other)const;
 
@@ -22,19 +22,19 @@ public:
 
 	Matrix33 Transpose();
 
-	void SetDiagonal(float x, float y, float z);
+	void SetDiagonal(double x, double y, double z);
 
 	void SetOrientation(const Quaternion& q);
 
-	float A() { return values[0]; };
-	float B() { return values[1]; };
-	float C() { return values[2]; };
-	float D() { return values[3]; };
-	float E() { return values[4]; };
-	float F() { return values[5]; };
-	float G() { return values[6]; };
-	float H() { return values[7]; };
-	float I() { return values[8]; };
+	double A() { return values[0]; };
+	double B() { return values[1]; };
+	double C() { return values[2]; };
+	double D() { return values[3]; };
+	double E() { return values[4]; };
+	double F() { return values[5]; };
+	double G() { return values[6]; };
+	double H() { return values[7]; };
+	double I() { return values[8]; };
